requirements.txt
openai
python-dotenv
openpyxl
jinja2
pydantic
python-docx
python-pptx
markdown-it-py
mdit-py-plugins
pillow
